split datastores to files.
